requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
beautifulsoup4==4.12.2
requests==2.31.0
openpyxl==3.1.2
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
python-decouple==3.8
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.6.0 